(ASC-221) Run test_cinder_service on dashboard_hosts
This commit moves the `test_cinder_service` into a `test_dashboard`
file that gathers hosts in the `dashboard_hosts` group to execute the
test on.  This test relies on the `cinder` command-line interface
which is not installed on the `cinder_volume` hosts.
